Extract light environment screen UI
