+++ service-worker.js
@@ -376,6 +376,7 @@ const APP_SHELL = [
   '/js/light-env.js',
   '/js/light-env-actions.js',
   '/js/light-env-store.js',
+  '/js/light-env-screen-ui.js',
   '/js/light-env-audits.js',
   '/js/light-env-ai-analysis.js',
   '/js/light-env-evening.js',
